Chat Network
Not sure if it will work. I was figure out the code for the chat network, because Unity has made many changes causing code on the internet to not work. The one here is something I found on the asset store, but when I click the play button nothing happens anyways so.(with and without the asset)